+++ float/src/ops.rs
@@ -1,5 +1,5 @@
 //! Re-exported relevant operator traits from `dashu-base`
 
-pub use dashu_base::math::EstimatedLog2;
-pub use dashu_base::ring::{CubicRoot, DivEuclid, DivRemEuclid, RemEuclid, SquareRoot};
+pub use dashu_base::math::{CubicRoot, EstimatedLog2, SquareRoot};
+pub use dashu_base::ring::{DivEuclid, DivRemEuclid, RemEuclid};
 pub use dashu_base::sign::Abs;